+++ src/controlers/userController.ts
@@ -1,6 +1,5 @@
 import { Request, Response } from "express";
 import { v4 } from "uuid";
-import users from "../db";
 import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient()
@@ -12,18 +11,18 @@ async function getAll(req:Request, res:Response){
 
 async function getById(req:Request, res:Response){
     const {id} : {id:string} = req.params as {id:string}
-    const user = await prisma.user.findUnique({where:{id:id}})
-    res.status(200).json(user);
+    if(await prisma.user.findUnique({where:{id:id}})){
+        const user = await prisma.user.findUnique({where:{id:id}})
+        res.status(200).json(user);
+     }
+     else{
+         res.status(404).json("Usuário não encontrado"); 
+     }
+    
 }
 
 async function addUser(req:Request, res:Response){
     const {name, username} = req.body as {name:string, username:string};
-    const newUser : User = {
-        id:v4(),
-        name,
-        username,
-        technologies: [] as Tecnology[]
-    }
     const user = await prisma.user.create({
         data: {
             id:v4(),
@@ -36,9 +35,21 @@ async function addUser(req:Request, res:Response){
 
 async function deleteUser(req:Request, res:Response){
     const {id} : {id:string} = req.params as {id:string}
-    const result = await prisma.user.delete({where:{id: id}})
-    console.log('deletou')
-    res.status(200).json(result);
+    if(await prisma.user.findUnique({where:{id:id}})){
+        const user = await prisma.user.findUnique({where:{id:id}})
+        const username = user?.username
+       if(!await prisma.tecnology.findFirst({where:{username:username}})){
+        const result = await prisma.user.delete({where:{id: id}})
+            res.status(200).json(result);
+       }
+       else{
+        res.status(401).json("Não foi possível excluir, pois existe ao menos uma tecnologia vinculada a esse usuário, por favor, exclua primeiro todas as tecnologias vinculadas.");
+       }
+    }
+    else{
+        res.status(404).json("Usuário não encontrado"); 
+    }
+    
 }
